T6/requirements.txt
antlr4-python3-runtime
svgwrite